Create index.php
Adding this code in order to make the truck list easier to work with. Currently I have to manually open the file to add or remove a truck. Adding this makes it possible to change it from different computers and well as for anyone else to add or remove a truck from the list.